pyproject.toml
[project]
name = "social-media-rag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pydantic",
    "streamlit",
    "matrix-nio[e2e]",
    "langchain-community",
    "pdf2image",
    "langchain",
    "langgraph",
    "langsmith",
    "langchain-milvus",
    "langchain-cohere",
    "langchain-google-genai",
    "pymilvus",
    "nest_asyncio",
    "aioboto3",
    "dependency-injector",
    "sqlmodel",
    "asyncpg",
]

[tool.setuptools.packages.find]
exclude = [
    "stores",
    "assets"
]
